+++ src/Enemy/KaronWingStateHack.cpp
@@ -18,7 +18,8 @@
 #include "Enemy/HackerDepthShadowMapCtrl.h"
 #include "Player/CollisionShapeInfo.h"
 #include "Player/CollisionShapeKeeper.h"
-#include "Player/HackerJudge.h"
+#include "Player/HackerJudgeNormalFall.h"
+#include "Player/HackerJudgeStartRun.h"
 #include "Player/HackerStateWingFly.h"
 #include "Player/PlayerActionTurnControl.h"
 #include "Player/PlayerCollider.h"
