+++ src/Util/HackerJudge.cpp
@@ -1,0 +1,5 @@
+#include "Util/HackerJudge.h"
+
+HackerJudge::HackerJudge(IUsePlayerHack** parent) {
+    mHacker = parent;
+};
